refactor: 拆分 workbench store 为 Zustand slices
- 提取 batch-slice.ts（批量分析逻辑，177 行）
- 提取 attachment-slice.ts（附件管理逻辑，76 行）
- 提取 streaming-slice.ts（流式消息/SSE 逻辑，216 行）
- workbench-store.ts 从 668 行精简到 ~130 行（session + model + agent）
- 使用 Zustand StateCreator slices 模式组合，公共 API 不变
